[not verified] Make store files easier to find,move to stores dir
