+++ projects/plugins/boost/app/features/image-guide/src/attach-guides.ts
@@ -1,4 +1,4 @@
-import { MeasurableImageStore } from './MeasurableImageStore';
+import { MeasurableImageStore } from './stores/MeasurableImageStore';
 import Main from './ui/Main.svelte';
 import type { MeasurableImage } from './MeasurableImage';
 import type { ImageComponentConfig } from './types';
