+++ projects/plugins/boost/app/features/image-guide/src/types.ts
@@ -1,4 +1,4 @@
-import type { MeasurableImageStore } from './MeasurableImageStore';
+import type { MeasurableImageStore } from './stores/MeasurableImageStore';
 import type { ComponentConstructorOptions } from 'svelte';
 
 export type MeasuredImage = {
